LayoutTests/http/tests/appcache/404-manifest-expected.txt

CONSOLE MESSAGE: Application Cache manifest could not be fetched, because the manifest had a 404 response.
Test that subresources can be loaded if manifest is not available. Should say SUCCESS.

SUCCESS
